+++ doc.go
@@ -1,0 +1,9 @@
+/*
+Package exporter provides all the building blocks
+of the prometheus-nats-exporter application.
+
+The main components are
+  - [EventListener]
+  - [EventHandler]
+*/
+package exporter