+++ event_handler_test.go
@@ -1,6 +1,7 @@
 package exporter
 
 import (
+	"encoding/json"
 	"testing"
 
 	"github.com/redis/go-redis/v9"
@@ -13,6 +14,11 @@ func Test_eventHandler_HandleEvent(t *testing.T) {
 
 	h := NewEventHandler(m, logger)
 
+	evt := event{
+		TaskName: "my-task",
+		TaskId:   "aabbccddee",
+	}
+
 	t.Run("malformed event", func(t *testing.T) {
 		msg := &redis.Message{Payload: "something"}
 
@@ -22,19 +28,45 @@ func Test_eventHandler_HandleEvent(t *testing.T) {
 		}
 	})
 
-	t.Run(string(Executing), func(t *testing.T) {
+	t.Run(string(executingEvent), func(t *testing.T) {
+		evt.Event = executingEvent
 
+		msg := evtToMessage(evt)
+
+		_ = h.HandleEvent(msg)
 	})
 
-	t.Run(string(Complete), func(t *testing.T) {
+	t.Run(string(completeEvent), func(t *testing.T) {
+		evt.Event = executingEvent
 
+		msg := evtToMessage(evt)
+
+		_ = h.HandleEvent(msg)
 	})
 
-	t.Run(string(Error), func(t *testing.T) {
+	t.Run(string(errorEvent), func(t *testing.T) {
+		evt.Event = executingEvent
 
+		msg := evtToMessage(evt)
+
+		_ = h.HandleEvent(msg)
 	})
 
-	t.Run(string(Locked), func(t *testing.T) {
+	t.Run(string(lockedEvent), func(t *testing.T) {
+		evt.Event = executingEvent
 
+		msg := evtToMessage(evt)
+
+		_ = h.HandleEvent(msg)
 	})
+}
+
+func evtToMessage(evt event) *redis.Message {
+	data, err := json.Marshal(evt)
+	if err != nil {
+		return nil
+	}
+	return &redis.Message{
+		Payload: string(data),
+	}
 }